requirements.txt
click
emoji
prometheus_client
python-telegram-bot
requests
telegram-click
